requirements.txt
aiogram>=3.4.0
telethon>=1.36.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
gspread>=5.12.0
google-auth>=2.25.0
google-genai>=1.0.0
anthropic>=0.42.0
apscheduler>=3.10.0
loguru>=0.7.0
rich>=13.7.0
aiohttp>=3.9.0
python-socks[asyncio]>=2.4.0
PySocks>=1.7.1
Pillow>=10.0.0
